Add coverAlbum component with dynamic canvas and SVG graphics
- Implemented a new React component `coverAlbum` that renders a visually rich cover layout.
- Utilized two canvas elements for layered text rendering with custom fonts and colors.
- Added SVG graphics for decorative elements to enhance the visual appeal.
- Included an image overlay with dynamic text positioning and line breaks for better readability.
- Applied a linear gradient background for aesthetic enhancement.
